requirements.txt
# JWT packages first to ensure they are installed first
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Core requirements
django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-q2==1.6.2
gunicorn==21.2.0
python-dotenv==1.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.10

# Document processing
bleach==6.0.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow>=9.0.0

# AWS S3 storage
boto3==1.34.34
django-storages==1.14.2

# AI services
pinecone-client==5.0.1
openai==1.54.4 